+++ WorkOrderDesk.Domain/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace WorkOrderDesk.Domain;
-
-public class Class1
-{
-
-}