+++ WorkOrderDesk.Domain/Users/User.cs
@@ -1,0 +1,6 @@
+namespace WorkOrderDesk.Domain.Users;
+
+public sealed class User
+{
+
+}