+++ WorkOrderDesk.Domain/WorkOrders/WorkOrderPriority.cs
@@ -1,0 +1,8 @@
+namespace workOrderDesk.Domain.WorkOrders;
+
+public enum WorkOrderPriority
+{
+    Low = 0,
+    Medium = 1,
+    Hight = 2,
+}