+++ WorkOrderDesk.Domain/WorkOrders/WorkOrderStatus.cs
@@ -1,0 +1,10 @@
+namespace WorkOrderDesk.Domain.WorkOrders;
+
+public enum WorkOrderStatus
+{
+    Backlog = 0,
+    Open = 1,
+    InProgress = 2,
+    Complete = 3,
+    Archived = 4
+}