+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## [Unreleased]
+## [Unreleased] 3/2/26
+
+### Added
+
+- WorkOrder domain model with status trasitions and audit time stamp
+- User entity + Userid
+
+## [Unreleased] 2/24/26
 
 ### Added